backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
supabase==2.18.1
python-dotenv==1.0.0

